+++ libs/ui-mobile/.storybook/preview.js
@@ -17,7 +17,7 @@ import '@ionic/react/css/text-alignment.css';
 import '@ionic/react/css/text-transformation.css';
 import '@ionic/react/css/flex-utils.css';
 import '@ionic/react/css/display.css';
-import theme from '../../ui-shared/src/lib/theme';
+import theme from '../../ui/src/styles/theme';
 import { ThemeProvider } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
