+++ libs/ui-shared/src/index.ts
@@ -1,2 +1,1 @@
-export * from './lib/theme';
-export * from './lib/client';
+export * from './lib/client/client';
