--- libs/ui-shared/src/lib/client/index.ts
@@ -1,6 +1,0 @@
-import { ApolloClient, InMemoryCache } from '@apollo/client';
-
-export const client = new ApolloClient({
-  uri: 'http://192.168.13.193:3333/graphql',
-  cache: new InMemoryCache(),
-});
