+++ libs/ui/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/ui';
 export * from './lib/support';
 export * from './lib/mocks';
 export * from './styles';
+export * from './styles/theme';
