+++ libs/ui/src/styles/index.jsx
@@ -1,3 +1,11 @@
+/**
+ * The global styles definitions
+ *
+ * Please note this file must stay as .jsx. Do not rename to .tsx,
+ * because it is used by the storybook.
+ *
+ * */
+
 import React from 'react';
 import { createGlobalStyle } from 'styled-components';
 import { Normalize } from 'styled-normalize';
