+++ nx.json
@@ -53,9 +53,6 @@
     },
     "ui-mobile": {
       "tags": []
-    },
-    "ui-shared": {
-      "tags": []
     }
   }
 }
