+++ package.json
@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "nx": "nx",
-    "start": "nx serve",
+    "start": "(set -a && source .env && nx serve && set +a)",
     "build": "nx build",
     "test": "nx test",
     "lint": "nx workspace-lint && nx lint",
